pyproject.toml
[tool.poetry]
name = "car-companion"
version = "0.1.0"
description = ""
authors = ["Mohammadali Azani <ali_azani@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "5.1.4"
psycopg2 = "^2.9.10"
django-environ = "^0.11.2"
django-debug-toolbar = "^4.4.6"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"
djoser = "^2.3.1"
social-auth-app-django = "^5.4.2"
drf-spectacular = "^0.28.0"
django-health-check = "^3.18.3"
django-unfold = "^0.43.0"
django-filter = "^24.3"
django-model-utils = "^5.0.0"
coverage = "^7.6.8"
django-colorfield = "^0.11.0"
django-guardian = "^2.4.0"
coverage-badge = "^1.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
